arqui3/services/s6/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
strawberry-graphql[fastapi]==0.216.1
pydantic==2.4.2
python-dotenv==1.0.0
alembic==1.12.1
aiokafka==0.8.1
asyncpg==0.29.0
python-multipart==0.0.6
python-dateutil==2.8.2 